Use common layout for all drawer + map pages
so that we only need to implement drawer open / close logic in one place.
